# Conflicts:
#	FTActors-be/src/main/resources/application.yml

+++ FTActors-be/src/main/java/com/a602/actors/domain/apply/service/ApplyServiceImpl.java
@@ -32,8 +32,12 @@ public class ApplyServiceImpl implements ApplyService {
     @Override
     @Transactional
     public void apply(ApplyDto applyDto) throws IOException {
-        String savedName = FileUtil.makeFileName(applyDto.getVideoFile().getOriginalFilename());
-        String url = FileUtil.uploadFile(applyDto.getVideoFile(), savedName, FolderType.APPLY_PATH);
+        String savedName ="";
+        String url = "";
+        if (applyDto.getVideoFile() != null) {
+            savedName = FileUtil.makeFileName(applyDto.getVideoFile().getOriginalFilename());
+            url = FileUtil.uploadFile(applyDto.getVideoFile(), savedName, FolderType.APPLY_PATH);
+        }
 
         Apply apply = Apply.builder()
                 .recruitment(recruitmentRepository.findById(applyDto.getRecruitmentId()).orElseThrow(() -> new RecruitmentException(ExceptionCodeSet.RECRUITMENT_NOT_FOUND)))
@@ -49,9 +53,11 @@ public class ApplyServiceImpl implements ApplyService {
     @Transactional
     public void applyCancel(Long recruitmentId, Long memberId) throws IOException {
         Apply apply = applyRepository.findByRecruitmentIdAndMemberId(recruitmentId, memberId);
-
+        if(apply ==null){
+            throw new ApplyException(ExceptionCodeSet.APPLY_NOT_FOUND);
+        }
         // 지원을 취소하기 전에 S3에서 파일을 삭제합니다.
-        String videoLink = apply.getVideoLink();
+        String videoLink = apply.getSavedName();
         if (videoLink != null) {
             FileUtil.deleteFile(videoLink, FolderType.APPLY_PATH);
         }
@@ -68,7 +74,7 @@ public class ApplyServiceImpl implements ApplyService {
                 .map(apply -> ApplyDto.builder()
                         .id(apply.getId())
                         .recruitmentId(apply.getRecruitment().getId())
-                        .memberId(apply.getMember().getId())
+                        .memberName(apply.getMember().getName())
                         .content(apply.getContent())
                         .videoLink(apply.getVideoLink())
                         .build())
@@ -82,7 +88,7 @@ public class ApplyServiceImpl implements ApplyService {
         ApplyDto applyDto = ApplyDto.builder()
                 .id(apply.getId())
                 .recruitmentId(apply.getRecruitment().getId())
-                .memberId(apply.getMember().getId())
+                .memberName(apply.getMember().getName())
                 .content(apply.getContent())
                 .videoLink(apply.getVideoLink())
                 .build();
@@ -96,14 +102,14 @@ public class ApplyServiceImpl implements ApplyService {
                 .map(apply -> ApplyDto.builder()
                         .id(apply.getId())
                         .recruitmentId(apply.getRecruitment().getId())
-                        .memberId(apply.getMember().getId())
+                        .memberName(apply.getMember().getName())
                         .content(apply.getContent())
                         .videoLink(apply.getVideoLink())
                         .build())
                 .collect(Collectors.toList());
         return applyDtoList;
     }
-    
+
     // 해당 공고에 유저가 지원했는지 확인
     @Override
     public int existApply(Long recruitmentId, Long memberId) {
